[demuxer] Fix seek param invalid issue

+++ dtplayer/dtplayer.c
@@ -360,8 +360,8 @@ int player_seekto(dtplayer_context_t * dtp_ctx, int seek_time)
     }
 
     player_state_t *play_stat = &dtp_ctx->state;
-    play_stat->cur_time = seek_time;
-    play_stat->cur_time_ms = seek_time * 1000;
+    play_stat->cur_time = (int64_t)seek_time;
+    play_stat->cur_time_ms = (int64_t)seek_time * 1000;
 
     set_player_status(dtp_ctx, PLAYER_STATUS_SEEK_ENTER);
     if (io_thread_running(dtp_ctx)) {
@@ -371,7 +371,7 @@ int player_seekto(dtplayer_context_t * dtp_ctx, int seek_time)
 
 
     int64_t start_time = (ctrl_info->first_time != -1) ? ctrl_info->first_time : ctrl_info->start_time;
-    int64_t target_time = seek_time * 1000000 + start_time;
+    int64_t target_time = (int64_t)seek_time * 1000000 + start_time;
     int ret = dtdemuxer_seekto(dtp_ctx->demuxer_priv, target_time);
     if (ret == -1) {
         goto FAIL;
